refactored some integration tests
