+++ tests/fixtures.py
@@ -34,7 +34,7 @@ from .fixtures_file_writer import GENERIC_START_RECORDING_COMMAND
 
 PATH_TO_CURRENT_FILE = get_current_file_abs_directory()
 QUEUE_CHECK_TIMEOUT_SECONDS = 1.1  # for is_queue_eventually_of_size, is_queue_eventually_not_empty, is_queue_eventually_empty, put_object_into_queue_and_raise_error_if_eventually_still_empty, etc. # Eli (10/28/20) issue encountered where even 0.5 seconds was insufficient, so raising to 1 second
-
+GENERIC_MAIN_LAUNCH_TIMEOUT_SECONDS = 15
 
 @pytest.fixture(scope="function", name="patched_shared_values_dict")
 def fixture_patched_shared_values_dict(mocker):
