+++ tests/test_main_integration.py
@@ -455,137 +455,6 @@ def test_system_states_and_recorded_metadata_with_update_to_file_writer_director
                     )
 
 
-@pytest.mark.slow
-@pytest.mark.timeout(INTEGRATION_TEST_TIMEOUT)
-@freeze_time(
-    datetime.datetime(
-        year=2020, month=7, day=16, hour=14, minute=19, second=55, microsecond=313309
-    )
-)
-def test_system_states_and_recording_files_with_file_directory_passed_in_cmd_line_args(
-    patched_xem_scripts_folder,
-    patched_firmware_folder,
-    test_process_manager_without_created_processes,
-    test_client,
-    fully_running_app_from_main_entrypoint,
-    mocker,
-):
-    expected_time = datetime.datetime(
-        year=2020, month=7, day=16, hour=14, minute=19, second=55, microsecond=313309
-    )
-    mocker.patch.object(
-        main,
-        "_get_timestamp_of_acquisition_sample_index_zero",
-        return_value=expected_time,
-    )
-    expected_timestamp = "2020_07_16_141955"
-
-    with tempfile.TemporaryDirectory() as expected_recordings_dir:
-        test_dict = {
-            "user_account_uuid": "0288efbc-7705-4946-8815-02701193f766",
-            "customer_account_uuid": "14b9294a-9efb-47dd-a06e-8247e982e196",
-            "recording_directory": expected_recordings_dir,
-        }
-        json_str = json.dumps(test_dict)
-        b64_encoded = base64.urlsafe_b64encode(json_str.encode("utf-8")).decode("utf-8")
-
-        command_line_args = [
-            "--skip-mantarray-boot-up",
-            f"--initial-base64-settings={b64_encoded}",
-        ]
-        _ = fully_running_app_from_main_entrypoint(command_line_args)
-        wait_for_subprocesses_to_start()
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[SERVER_READY_STATE]
-        )
-
-        response = requests.get(f"{get_api_endpoint()}boot_up")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[INSTRUMENT_INITIALIZING_STATE]
-        )
-        assert system_state_eventually_equals(CALIBRATION_NEEDED_STATE, 3) is True
-
-        response = requests.get(f"{get_api_endpoint()}start_calibration")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[CALIBRATING_STATE]
-        )
-        assert (
-            system_state_eventually_equals(CALIBRATED_STATE, CALIBRATED_WAIT_TIME)
-            is True
-        )
-
-        response = requests.get(f"{get_api_endpoint()}start_managed_acquisition")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[BUFFERING_STATE]
-        )
-        assert (
-            system_state_eventually_equals(
-                LIVE_VIEW_ACTIVE_STATE, LIVE_VIEW_ACTIVE_WAIT_TIME
-            )
-            is True
-        )
-
-        expected_barcode = GENERIC_START_RECORDING_COMMAND[
-            "metadata_to_copy_onto_main_file_attributes"
-        ][PLATE_BARCODE_UUID]
-        response = requests.get(
-            f"{get_api_endpoint()}start_recording?barcode={expected_barcode}&is_hardware_test_recording=False"
-        )
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[RECORDING_STATE]
-        )
-
-        time.sleep(3)  # Tanner (6/15/20): This allows data to be written to files
-
-        response = requests.get(f"{get_api_endpoint()}stop_recording")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert response.json()["ui_status_code"] == str(
-            SYSTEM_STATUS_UUIDS[LIVE_VIEW_ACTIVE_STATE]
-        )
-
-        response = requests.get(f"{get_api_endpoint()}stop_managed_acquisition")
-        assert response.status_code == 200
-        response = requests.get(f"{get_api_endpoint()}system_status")
-        assert response.status_code == 200
-        assert (
-            system_state_eventually_equals(
-                CALIBRATED_STATE, STOP_MANAGED_ACQUISITION_WAIT_TIME
-            )
-            is True
-        )
-
-        test_process_manager_without_created_processes.soft_stop_processes()
-
-        fw_process = (
-            test_process_manager_without_created_processes.get_file_writer_process()
-        )
-        fw_process.close_all_files()
-        actual_set_of_files = set(
-            os.listdir(
-                os.path.join(
-                    expected_recordings_dir, f"{expected_barcode}__{expected_timestamp}"
-                )
-            )
-        )
-        assert len(actual_set_of_files) == 24
-
-        test_process_manager_without_created_processes.hard_stop_and_join_processes()
 
 
 @pytest.mark.slow
